app/requirements.txt
uvicorn
fastapi
sqlalchemy
psycopg2
python-dotenv
schedule